+++ packages/stream_chat/lib/src/core/models/channel_config.dart
@@ -24,6 +24,7 @@ class ChannelConfig {
     this.typingEvents = false,
     this.uploads = false,
     this.urlEnrichment = false,
+    this.skipLastMsgUpdateForSystemMsgs = false,
   })  : createdAt = createdAt ?? DateTime.now(),
         updatedAt = updatedAt ?? DateTime.now();
 
@@ -78,6 +79,13 @@ class ChannelConfig {
 
   /// True if urls appears as attachments
   final bool urlEnrichment;
+
+  /// If true the last message at date will not be updated when a system message
+  /// is added.
+  ///
+  /// This is useful for scenarios where you want to track the last time a user
+  /// message was added to the channel.
+  final bool skipLastMsgUpdateForSystemMsgs;
 
   /// Serialize to json
   Map<String, dynamic> toJson() => _$ChannelConfigToJson(this);