+++ packages/stream_chat_persistence/CHANGELOG.md
@@ -1,8 +1,7 @@
 ## Upcoming
 
-✅ Added
-
-- Add member.extraData field
+- Added support for `Message.restrictedVisibility` field.
+- Added support for `Member.extraData` field.
 
 ## 9.4.0